refactor: move out fairly general color utility
